# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Electrical Shop Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: lowStock
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SALES ───────────────────────────────────────────────────────────────────
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: customerId
          schema:
            type: integer
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CUSTOMERS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        costPrice:
          type: number
        sellingPrice:
          type: number
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        unit:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, costPrice, sellingPrice, stockQuantity, lowStockThreshold, unit, createdAt, updatedAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        costPrice:
          type: number
        sellingPrice:
          type: number
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
          default: 5
        unit:
          type: string
          default: "pcs"
      required: [name, costPrice, sellingPrice, stockQuantity]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        costPrice:
          type: number
        sellingPrice:
          type: number
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        unit:
          type: string
    Sale:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        totalAmount:
          type: number
        paidAmount:
          type: number
        creditAmount:
          type: number
        notes:
          type: string
          nullable: true
        source:
          type: string
          enum: [web, telegram]
        createdAt:
          type: string
          format: date-time
      required: [id, items, totalAmount, paidAmount, creditAmount, source, createdAt]
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
          nullable: true
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        subtotal:
          type: number
      required: [productName, quantity, unitPrice, subtotal]
    CreateSaleBody:
      type: object
      properties:
        customerId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateSaleItemBody"
        paidAmount:
          type: number
        notes:
          type: string
          nullable: true
        source:
          type: string
          enum: [web, telegram]
          default: web
      required: [items, paidAmount]
    CreateSaleItemBody:
      type: object
      properties:
        productId:
          type: integer
          nullable: true
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
      required: [productName, quantity, unitPrice]
...