# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Electrical Shop Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  description?: string | null;
  sku?: string | null;
  category?: string | null;
  costPrice: number;
  sellingPrice: number;
  stockQuantity: number;
  lowStockThreshold?: number;
  unit?: string;
}