# lib/api-zod/src/generated/types/createSaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Electrical Shop Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSaleBodySource } from "./createSaleBodySource";
import type { CreateSaleItemBody } from "./createSaleItemBody";

export interface CreateSaleBody {
  customerId?: number | null;
  items: CreateSaleItemBody[];
  paidAmount: number;
  notes?: string | null;
  source?: CreateSaleBodySource;
}